+++ src/components/Scn/components/ScnElement/ScnElement.tsx
@@ -16,6 +16,7 @@ import { ScnLink } from '../ScnLink';
 import { Struct, StyledScg, StyledSpinner, StyledSwitchScgScn } from '../ScStruct/styled';
 
 import { Arc, Child, LinkedNodes, Marker, Modifier, RightSide, StyledLinkedNode, Wrapper } from './styled';
+import { getRandomInt } from '@utils/getRandomInt';
 
 interface IProps {
   tree: IScnNode;
@@ -137,7 +138,7 @@ const ScnElementWrapper = ({ tree, isRoot = false }: IProps) => {
                 <Modifier>
                   {modifiers.map((modifier) => (
                     <Fragment key={`${arc.addr}${modifier.addr}`}>
-                      <ScnLink addr={modifier.addr} />
+                      <ScnLink addr={modifier.addr} loaderHeight={18} loaderWidth={`${getRandomInt(20, 60)}%`} />
                       <ModifierArc type={modifier.modifierArcs[0].type} />
                     </Fragment>
                   ))}