+++ docs/changelog.md
@@ -83,3 +83,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - export InputStatus
+- Change style of skeleton of SCn